Move Icon component to its own file
